consumer/requirements.txt
kafka-python>=2.0.2
scikit-learn>=1.2.0
pandas>=1.5.2
nltk>=3.8
